Fixed command filtering
Previously a filter requiring all commands end with "Immediate" was
added - which is removed in this commit. Additionally, there was a error
thrown when trying to read control/indicator data for SubSystems.vi, so
it was manually excluded. I also added template vi (.vit) filtering
which according to some comment already existed, but I could find no
such evidence to that claim.