Initial bezier spline code
Spline generation has not been throughly tested, although I did write a
vi to test it
I'm not really sure what the bezier spline typedef will look like later,
so I just made it an array of bezier curves. Because I didnt make it a
struct adding more meta-data in the future will be a pain.