Chebyhell
The NI Cheby stuff couldnt be modified for my needs as it mainly was for
fitting functions and not a set of 2d points. I need to fit 2d functions
to easily invert polynomial as for points this is easy (swap x and y),
but for abritrary functions (bezier is as good as arbitrary as far as im
concerned), with the x value requirement in the NI code, its not very
doable.

Chebfit and Chebeval2 are the good vi's, the rest will be removed in a
later commit.

Chebfit solves for the coefficents for the associated vandermonde
function for chebyshev polynomials